+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#8](https://github.com/LaxarJS/ax-input-control/issues/8): tooltip: allow to set on parent element
 - [#7](https://github.com/LaxarJS/ax-input-control/issues/7): validation: fixed setting custom validation message